data_collector/requirements.txt
berserk>=0.13.0
python-dotenv>=1.0.0
